requirements.txt
python-barcode==0.15.1
pillow==10.3.0
